# api_schema/openapi-schema.json
{
    "swagger": "2.0",
    "info": {
        "title": "SMERP API",
        "description": "Auto-generated docs for SMERP API",
        "version": "v1"
    },
    "basePath": "/api/v1",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "Basic": {
            "type": "basic"
        }
    },
    "security": [
        {
            "Basic": []
        }
    ],
    "paths": {
...
        "/material-demands/{id}/delete/": {
            "post": {
                "operationId": "material-demands_delete",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MaterialDemand"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/MaterialDemand"
                        }
                    }
                },
                "tags": [
                    "material-demands"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this material demand.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
        "/material-demands/{id}/recover/": {
            "post": {
                "operationId": "material-demands_recover",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MaterialDemand"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/MaterialDemand"
                        }
                    }
                },
                "tags": [
                    "material-demands"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this material demand.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
...
        "/materials/{id}/delete/": {
            "post": {
                "operationId": "materials_delete",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Material"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Material"
                        }
                    }
                },
                "tags": [
                    "materials"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this material.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
        "/materials/{id}/recover/": {
            "post": {
                "operationId": "materials_recover",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Material"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Material"
                        }
                    }
                },
                "tags": [
                    "materials"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this material.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
...
        "/procurement-order-lines/{id}/delete/": {
            "post": {
                "operationId": "procurement-order-lines_delete",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ProcurementOrderLine"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/ProcurementOrderLine"
                        }
                    }
                },
                "tags": [
                    "procurement-order-lines"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this procurement order line.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
        "/procurement-order-lines/{id}/recover/": {
            "post": {
                "operationId": "procurement-order-lines_recover",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ProcurementOrderLine"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/ProcurementOrderLine"
                        }
                    }
                },
                "tags": [
                    "procurement-order-lines"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this procurement order line.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
...
        "/procurement-orders/{id}/delete/": {
            "post": {
                "operationId": "procurement-orders_delete",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ProcurementOrder"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/ProcurementOrder"
                        }
                    }
                },
                "tags": [
                    "procurement-orders"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this procurement order.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
        "/procurement-orders/{id}/recover/": {
            "post": {
                "operationId": "procurement-orders_recover",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ProcurementOrder"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/ProcurementOrder"
                        }
                    }
                },
                "tags": [
                    "procurement-orders"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this procurement order.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
...
    },
    "definitions": {
...
        "MaterialDemand": {
            "required": [
                "material",
                "quantity",
                "deadline",
                "description"
            ],
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "demand_no": {
                    "title": "Talep No",
                    "type": "string",
                    "readOnly": true,
                    "minLength": 1,
                    "x-nullable": true
                },
                "material": {
                    "title": "Malzeme",
                    "type": "integer"
                },
                "quantity": {
                    "title": "Miktar",
                    "type": "string",
                    "format": "decimal"
                },
                "uom": {
                    "title": "Birim",
                    "type": "string",
                    "enum": [
                        "ADT",
                        "KG",
                        "G",
                        "L",
                        "ML",
                        "M",
                        "BOX",
                        "PLT"
                    ]
                },
                "deadline": {
                    "title": "Son Tarih",
                    "type": "string",
                    "format": "date"
                },
                "description": {
                    "title": "Açıklama",
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1
                },
                "status": {
                    "title": "Durum",
                    "type": "string",
                    "enum": [
                        "submitted",
                        "approved",
                        "closed"
                    ],
                    "readOnly": true
                },
                "created_by": {
                    "title": "Created by",
                    "type": "string",
                    "readOnly": true
                },
                "created_at": {
                    "title": "Created at",
                    "type": "string",
                    "readOnly": true
                }
            }
        },
        "Material": {
            "required": [
                "description"
            ],
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "name": {
                    "title": "Ad",
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1,
                    "x-nullable": true
                },
                "category": {
                    "title": "Kategori",
                    "type": "string",
                    "enum": [
                        "supplied",
                        "cardboard",
                        "part",
                        "good",
                        "administrative",
                        "pallet",
                        "undefined"
                    ]
                },
                "internal_code": {
                    "title": "İç Kod",
                    "type": "string",
                    "readOnly": true,
                    "minLength": 1,
                    "x-nullable": true
                },
                "description": {
                    "title": "Açıklama",
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1
                },
                "created_by": {
                    "title": "Created by",
                    "type": "string",
                    "readOnly": true
                },
                "created_at": {
                    "title": "Created at",
                    "type": "string",
                    "readOnly": true
                }
            }
        },
        "ProcurementOrderLine": {
            "required": [
                "material",
                "quantity",
                "quantity_received"
            ],
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "po": {
                    "title": "Po",
                    "type": "integer"
                },
                "material": {
                    "title": "Malzeme",
                    "type": "integer"
                },
                "uom": {
                    "title": "Birim",
                    "type": "string",
                    "enum": [
                        "ADT",
                        "KG",
                        "G",
                        "L",
                        "ML",
                        "M",
                        "BOX",
                        "PLT"
                    ]
                },
                "quantity": {
                    "title": "Quantity",
                    "type": "number",
                    "format": "decimal"
                },
                "quantity_received": {
                    "title": "Quantity received",
                    "type": "number",
                    "format": "decimal"
                },
                "unit_price": {
                    "title": "Unit price",
                    "type": "number",
                    "format": "decimal",
                    "x-nullable": true
                },
                "tax_rate": {
                    "title": "Tax rate",
                    "type": "number",
                    "format": "decimal",
                    "x-nullable": true
                },
                "quantity_left": {
                    "title": "Quantity left",
                    "type": "string",
                    "readOnly": true
                },
                "total_with_tax": {
                    "title": "Total with tax",
                    "type": "string",
                    "readOnly": true
                },
                "total_without_tax": {
                    "title": "Total without tax",
                    "type": "string",
                    "readOnly": true
                },
                "created_by": {
                    "title": "Created by",
                    "type": "string",
                    "readOnly": true
                },
                "created_at": {
                    "title": "Created at",
                    "type": "string",
                    "readOnly": true
                }
            }
        },
        "ProcurementOrder": {
            "required": [
                "payment_term",
                "payment_method",
                "incoterms",
                "due_in_days",
                "description"
            ],
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "payment_term": {
                    "title": "Ödeme Koşulları",
                    "type": "string",
                    "enum": [
                        "NET_T",
                        "DUE_ON_RECEIPT",
                        "CIA",
                        "CWO",
                        "COD",
                        "CAD",
                        "EOM",
                        "T_EOM",
                        "PROX",
                        "PARTIAL_ADVANCE",
                        "LC",
                        "DC",
                        "X_Y_NET_T",
                        "TRADE_DISCOUNT"
                    ]
                },
                "payment_method": {
                    "title": "Ödeme Yöntemi",
                    "type": "string",
                    "enum": [
                        "BANK_TRANSFER",
                        "CASH",
                        "CREDIT_CARD",
                        "CHEQUE",
                        "EFT",
                        "PROMISSORY_NOTE",
                        "OTHER"
                    ]
                },
                "incoterms": {
                    "title": "Teslim Şekli (Incoterms)",
                    "type": "string",
                    "enum": [
                        "EXW",
                        "FCA",
                        "FAS",
                        "FOB",
                        "CFR",
                        "CIF",
                        "CPT",
                        "CIP",
                        "DAP",
                        "DPU",
                        "DDP"
                    ]
                },
                "trade_discount": {
                    "title": "Ticari İskonto",
                    "type": "string",
                    "format": "decimal"
                },
                "due_in_days": {
                    "title": "Vade Günü",
                    "type": "string"
                },
                "due_discount": {
                    "title": "Vade İskontosu",
                    "type": "string",
                    "format": "decimal"
                },
                "due_discount_days": {
                    "title": "Vade İskonto Günü",
                    "type": "string",
                    "x-nullable": true
                },
                "invoice_accepted": {
                    "title": "Invoice accepted",
                    "type": "string",
                    "readOnly": true
                },
                "invoice_date": {
                    "title": "Fatura Tebliğ Tarihi",
                    "type": "string",
                    "format": "date",
                    "x-nullable": true
                },
                "last_payment_date": {
                    "title": "Last payment date",
                    "type": "string",
                    "readOnly": true
                },
                "description": {
                    "title": "Açıklama",
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1
                },
                "status": {
                    "title": "Durum",
                    "type": "string",
                    "enum": [
                        "draft",
                        "submitted",
                        "approved",
                        "rejected",
                        "ordered",
                        "billed",
                        "paid",
                        "cancelled"
                    ]
                },
                "currency": {
                    "title": "Currency",
                    "type": "string",
                    "enum": [
                        "TRY",
                        "USD",
                        "EUR",
                        "GBP",
                        "JPY",
                        "CHF",
                        "RUB",
                        "CNY",
                        "SEK",
                        "AUD"
                    ]
                },
                "delivery_address": {
                    "title": "Teslimat Adresi",
                    "type": "string",
                    "maxLength": 250,
                    "minLength": 1,
                    "x-nullable": true
                },
                "created_by": {
                    "title": "Created by",
                    "type": "string",
                    "readOnly": true
                },
                "created_at": {
                    "title": "Created at",
                    "type": "string",
                    "readOnly": true
                },
                "total_price_without_tax": {
                    "title": "Total price without tax",
                    "type": "string",
                    "readOnly": true
                },
                "total_price_with_tax": {
                    "title": "Total price with tax",
                    "type": "string",
                    "readOnly": true
                },
                "all_received": {
                    "title": "All received",
                    "type": "string",
                    "readOnly": true
                },
                "lines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ProcurementOrderLine"
                    }
                },
                "vendor": {
                    "title": "Vendor",
                    "type": "integer"
                }
            }
        },
...
    }
}